+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name="1.7.2"></a>
+## [1.7.2](https://github.com/gabyvs/ng2-ue-utils/compare/1.7.1...1.7.2) (2016-10-13)
+
+### Improvements
+
+* **notification:** A class icon prefix can be used. This allows to use _Fontawesome_ instead of _glyphicons_.
+
+### Bug Fixes
+
+* **notification:** Coming soon message was removed.
+
 <a name="1.7.1"></a>
 ## [1.7.1](https://github.com/gabyvs/ng2-ue-utils/compare/1.7.0...1.7.1) (2016-10-06)